Load the dbd static library using ctypes PyDLL convention instead of CDLL
The difference is that this way we hold onto the Python GIL.  Mostly this
makes no difference, as these are very quick function calls, but it turns
out that if there is another Python thread running then contention for the
GIL can wreck performance here.

The ctypes documentation is not particularly helpful, saying in particular
"this is only useful to call Python C api functions directly", which
doesn't seem to be correct.